requirements.txt
flask~=2.2.2
tensorflow~=2.17.0
numpy~=1.26.4
PIL
Pillow~=11.0.0
requests~=2.32.3
opencv-python~=4.10.0.84
APScheduler~=3.10.4
matplotlib~=3.9.2
python-dotenv~=1.0.1